Text.Seonbi.Facade

Description

Provides higher-level APIs. Read transformHtmlText function first, and then see also Configuration type.

HTML transformation

transformHtmlText :: Monad m => Configuration m a -> Text -> m Text

Transforms a given HTML text. Nothing if it fails to parse the given HTML text.

transformHtmlLazyText :: Monad m => Configuration m a -> Text -> m Text

A lazy version of transformHtmlText function.

Configuration and presets

data Monad m => Configuration m a

Transformation settings. For the most cases, you could use one of presets:

ko_KR

ko_KP

Constructors

Configuration

Fields

debugLogger :: Maybe (HtmlEntity -> m a)

An optional debugging logger to print its internal AST.

xhtml :: Bool

Whether to take and result in XHTML instead of HTML.

quote :: Maybe QuoteOption

An option to decide how quotation marks are rendered. If Nothing no quotes are transformed.

cite :: Maybe CiteOption

An option to transform folk-citing quotes (e.g., <<한겨레>>) into proper citing quotes (e.g., 《한겨레》).

arrow :: Maybe ArrowOption

Settings to transform arrow-looking punctuations into proper arrows. If Nothing no arrows are transformed.

ellipsis :: Bool

Whether to transform triple periods into a proper ellipsis.

emDash :: Bool

Whether to transform folk em dashes into proper em dashes.

stop :: Maybe StopOption

Settings to normalize stops (periods, commas, and interpuncts). If Nothing stops are never touched.

hanja :: Maybe HanjaOption

Settings to deal with Sino-Korean words.

ko_KP :: Monad m => Configuration m a

Preset Configuration for North Korean orthography.

ko_KR :: Monad m => Configuration m a

Preset Configuration for South Korean orthography.

presets :: (Ord k, IsString k, Monad m) => Map k (Configuration m a)

A mapping of locale code strings (e.g., "ko-kr") to the corresponding Configuration presets (e.g., ko_KR).

Dictionaries

type HanjaDictionary = Trie Text

Represents a dictionary that has hanja keys and values of their corresponding hangul readings, e.g., [("敗北", "패배")].

readDictionaryFile :: FilePath -> IO HanjaDictionary

Loads a dictionary file. The file consists of two-column TSV (tab-separated values); the first column is hanja and the second column is hangul.

southKoreanDictionary :: IO HanjaDictionary

Loads Standard Korean Language Dictionary (標準國語大辭典) data.

Options

data ArrowOption

Settings to transform arrow-looking punctuations into proper arrows.

Constructors

ArrowOption

Fields

bidirArrow :: Bool

Whether to transform bi-directional arrows as well as left/rightwards arrows.

doubleArrow :: Bool

Whether to transform double arrows as well as single arrows.

data CiteOption

An option to transform folk-citing quotes (e.g., <<한겨레>>) into proper citing quotes (e.g., 《한겨레》).

Constructors

AngleQuotes
Cite a title using angle quotes, used by South Korean orthography in horizontal writing (橫書), e.g., 《나비와 엉겅퀴》 or 〈枾崎의 바다〉.

CornerBrackets
Cite a title using corner brackets, used by South Korean orthography in vertical writing (縱書) and Japanese orthography, e.g., 『나비와 엉겅퀴』 or 「枾崎의 바다」.

AngleQuotesWithCite
Same as AngleQuotes except it wraps the title with a <cite> tag.

CornerBracketsWithCite
Same as CornerBrackets except it wraps the title with a <cite> tag.

data HanjaOption

Settings to deal with Sino-Korean words.

Constructors

HanjaOption

Fields

rendering :: HanjaRenderingOption

How to render Sino-Korean words.

reading :: HanjaReadingOption

How to rewrite Sino-Korean words in hangul.

data HanjaReadingOption

Settings to read Sino-Korean words.

Constructors

HanjaReadingOption

Fields

initialSoundLaw :: Bool

Whether to apply Initial Sound Law (頭音法則) or not.

dictionary :: HanjaDictionary

A dictionary which has hanja readings. Keys are hanja words and values are their corresponding hangul readings, e.g.:

[("敗北", "패배"), ("北極", "북극")] :: HanjaDictionary

data HanjaRenderingOption

Available options to render Sino-Korean words.

Constructors

HangulOnly
Renders a word in hangul-only, no hanja at all (e.g., 안녕히).

HanjaInParentheses
Renders a word in hangul followed by hanja in parentheses (e.g., 안녕(安寧)히).

DisambiguatingHanjaInParentheses
Renders words in hangul-only for the most part, and if there are homophones in a document put their hanja notation in parentheses (e.g., 안녕히 or 소수(小數)와 소수(素數)).

HanjaInRuby
Renders a word in ruby tag (e.g., <ruby>安寧<rp>(</rp><rt>안녕</rt><rp>)</rp></ruby>히).

Please read Use Cases & Exploratory Approaches for Ruby Markup as well for more information.

data QuoteOption

An option to decide how quotation marks are rendered.

Constructors

CurvedQuotes
English-style curved quotes (‘: U+2018, ’: U+2019, “: U+201C, ”: U+201D), which are used by South Korean orthography.

VerticalCornerBrackets
Vertical corner brackets (﹁: U+FE41, ﹂: U+FE42, ﹃: U+FE43, ﹄: U+FE44), which are used by East Asian orthography.

HorizontalCornerBrackets
Traditional horizontal corner brackets (「: U+300C, 」: U+300D, 『: U+300E, 』: U+300F), which are used by East Asian orthography.

Guillemets
East Asian guillemets (〈: U+3008, 〉: U+3009, 《: U+300A, 》: U+300B), which are used by North Korean orthography.

CurvedSingleQuotesWithQ
Use English-style curved quotes (‘: U+2018, ’: U+2019) for single quotes, and HTML <q> tags for double quotes.

VerticalCornerBracketsWithQ
Use vertical corner brackets (﹁: U+FE41, ﹂: U+FE42) for single quotes, and HTML <q> tags for double quotes.

HorizontalCornerBracketsWithQ
Use horizontal corner brackets (「: U+300C, 」: U+300D) for single quotes, and HTML <q> tags for double quotes.

data StopOption

Settings to normalize stops (periods, commas, and interpuncts) in docs.

Constructors

Horizontal
Stop sentences in the modern Korean style which follows Western stops. E.g.:

봄·여름·가을·겨울. 어제, 오늘.

HorizontalWithSlashes
Similar to horizontalStops except slashes are used instead of interpuncts. E.g.:

봄/여름/가을/겨울. 어제, 오늘.

Vertical
Stop sentences in the pre-modern Korean style which follows Chinese stops. E.g.:

봄·여름·가을·겨울。어제、오늘。