Text.Seonbi.Hangul

Documentation

type JamoTriple = (Char, Char, Maybe Char)

A triple of an initial consonant, a vowel, and an optional final consonant.

fromJamoTriple :: JamoTriple -> Maybe Char

Composes hangul jamo triple into a hangul syllable.

>>> fromJamoTriple ('ᄀ', 'ᅡ', Nothing)
Just '가'
>>> fromJamoTriple ('ᄀ', 'ᅳ', Just 'ᆯ')
Just '글'

isHangulSyllable :: Char -> Bool

Checks if a character is a hangul letter and a complete syllable.

>>> isHangulSyllable '가'
True
>>> isHangulSyllable 'ㄱ'
False

toJamoTriple :: Char -> Maybe JamoTriple

Takes a complete hangul syllable apart into consonants and a vowel. Returns Nothing for non-hangul letters.

>>> toJamoTriple '가'
Just ('ᄀ','ᅡ',Nothing)
>>> toJamoTriple '글'
Just ('ᄀ','ᅳ',Just 'ᆯ')
>>> toJamoTriple 'A'
Nothing