Text.Seonbi.Html

Description

Since Seonbi's primitive unit to transform is HTML, this module deals with HTML.

HTML scanner

See more on Text.Seonbi.Html.Scanner module.

data Result r

Constructors

Fail Text [String] String

Done Text r

scanHtml :: Text -> Result [HtmlEntity]

HTML printer

See more on Text.Seonbi.Html.Printer module.

printHtml :: [HtmlEntity] -> Text

Print the list of HtmlEntity into a lazy Text.

>>> :set -XOverloadedStrings
>>> import Text.Seonbi.Html.Scanner
>>> let Done "" tokens = scanHtml "<p>Hello,<br>\n<em>world</em>!</p>"
>>> printHtml tokens
"<p>Hello,<br>\n<em>world</em>!</p>"

printXhtml :: [HtmlEntity] -> Text

Similar to printHtml except it renders void (self-closing) tags as like br/ instead of br.

>>> :set -XOverloadedStrings
>>> import Text.Seonbi.Html.Scanner
>>> let Done "" tokens = scanHtml "<p>Hello,<br>\n<em>world</em>!</p>"
>>> printXhtml tokens
"<p>Hello,<br/>\n<em>world</em>!</p>"

Note that normal tags are not rendered as self-closed; only void tags according to HTML specification are:

>>> let Done "" tokens' = scanHtml "<p></p><p><br></p>"
>>> printXhtml tokens'
"<p></p><p><br/></p>"

HTML entities

See more on Text.Seonbi.Html.Entity module.

data HtmlEntity

An event entity emitted by scanHtml.

Constructors

HtmlStartTag
Represent a token which opens an HTML element.

Note that rawAttributes is not a parsed and structured data but a raw string as its name implies.

The tagStack doesn't include the corresponding opened tag.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

tag :: HtmlTag

rawAttributes :: HtmlRawAttrs

HtmlEndTag
Represent a token which closes an HTML element. The tagStack doesn't include the corresponding closed tag.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

tag :: HtmlTag

HtmlText
Represent a token of a text node. Note that rawText is not a parsed and structured data but a raw string as its name implies. There can be continuously more than one HtmlText values can be emitted even if they are not separated by element openings or closings.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

rawText :: Text

HtmlCdata
Represent a token of a CDATA section.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

text :: Text

HtmlComment
Represent a token of an HTML comment.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

comment :: Text

type HtmlRawAttrs = Text

HTML tags

See more on Text.Seonbi.Html.Tag module.

data HtmlTag

HTML tags. This enumeration type contains both HTML 5 and 4 tags for maximum compatibility.

Constructors

A

Abbr

Acronym

Address

Area

Article

Aside

Audio

B

Base

Bdi

Bdo

Big

BlockQuote

Body

BR

Button

Canvas

Caption

Cite

Code

Col

ColGroup

Data

DataList

DD

Del

Details

Dfn

Dialog

Div

DL

DT

Em

Embed

FieldSet

FigCaption

Figure

Footer

Form

H1

H2

H3

H4

H5

H6

Head

Header

HR

Html

I

IFrame

Img

Input

Ins

Kbd

Label

Legend

LI

Link

Main

Map

Mark

Meta

Meter

Nav

NoScript

Object

OL

OptGroup

Option

Output

P

Param

Picture

Pre

Progress

Q

RB

RP

RT

RTC

Ruby

S

Samp

Script

Select

Section

Small

Source

Span

Strong

Style

Sub

Summary

Sup

Table

TBody

TD

Template

TFoot

TextArea

TH

THead

Time

Title

TR

Track

TT

U

UL

Var

Video

WBR

data HtmlTagKind

The six kinds of HTML elements.

Constructors

Void

Template'

RawText

EscapableRawText

Foreign

Normal

htmlTagKind :: HtmlTag -> HtmlTagKind

The kind of an HtmlTag.

>>> Data.Set.filter ((== EscapableRawText) . htmlTagKind) htmlTags
fromList [TextArea,Title]

htmlTagName :: HtmlTag -> Text

The name of an HtmlTag in lowercase.

>>> htmlTagName TextArea
"textarea"

\ t -> htmlTagName t == (toLower $ pack $ show (t :: HtmlTag))

HTML text normalization

See more on Text.Seonbi.Html.TextNormalizer module.

normalizeText :: [HtmlEntity] -> [HtmlEntity]

As scanHtml may emit two or more continuous HtmlText fragments even if these can be represented as only one HtmlText fragment, it makes postprocessing hard.

The normalizeText function concatenates such continuous HtmlText fragments into one if possible so that postprocessing can be easy:

>>> :set -XOverloadedStrings -XOverloadedLists
>>> normalizeText [HtmlText [] "Hello, ", HtmlText [] "world!"]
[HtmlText {tagStack = fromList [], rawText = "Hello, world!"}]

It also transforms all HtmlCdata fragments into an HtmlText together.

>>> :{
normalizeText [ HtmlText [] "foo "
              , HtmlCdata [] "<bar>", HtmlText [] " baz!"
              ]
:}
[HtmlText {tagStack = fromList [], rawText = "foo &lt;bar&gt; baz!"}]

HTML hierarchical stacks

See more on Text.Seonbi.Html.TagStack module.

data HtmlTagStack

Represents a hierarchy of a currently parsing position in an HtmlTag tree.

For example, if an scanHtml has read "<a href="#"><b><i>foo</i> bar" it is represented as HtmlTagStack [B, A].

Note that the tags are stored in reverse order, from the deepest to the shallowest, to make inserting a more deeper tag efficient.