Text.Seonbi.Html.Entity

Documentation

data HtmlEntity

An event entity emitted by scanHtml.

Constructors

HtmlStartTag
Represent a token which opens an HTML element.

Note that rawAttributes is not a parsed and structured data but a raw string as its name implies.

The tagStack doesn't include the corresponding opened tag.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

tag :: HtmlTag

rawAttributes :: HtmlRawAttrs

HtmlEndTag
Represent a token which closes an HTML element. The tagStack doesn't include the corresponding closed tag.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

tag :: HtmlTag

HtmlText
Represent a token of a text node. Note that rawText is not a parsed and structured data but a raw string as its name implies. There can be continuously more than one HtmlText values can be emitted even if they are not separated by element openings or closings.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

rawText :: Text

HtmlCdata
Represent a token of a CDATA section.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

text :: Text

HtmlComment
Represent a token of an HTML comment.

Fields

tagStack :: HtmlTagStack

A stack of HtmlTags that represents a hierarchy of a currently parsing position in an HtmlTag tree.

comment :: Text

type HtmlRawAttrs = Text