Text.Seonbi.Html.Lang

Documentation

data LangHtmlEntity

Annotates HtmlEntity with the lang tag extracted from it or its ancestors.

Constructors

LangHtmlEntity

Fields

lang :: Maybe LanguageTag

The lang tag extracted from the HTML entity or its ancestors.

entity :: HtmlEntity

The annotated HTML entity.

type LanguageTag = Text

Represents a language tag. Although it is defined as an alias for Text, it can be structured in the future. Do not use its contents directly.

annotateWithLang :: [HtmlEntity] -> [LangHtmlEntity]

Annotates the given HTML entities with the language tag extracted from their lang attributes. If a parent entity has lang attribute, its all descendants are annotated with the same language tag.

extractLang

Arguments

:: HtmlRawAttrs
A raw HTML attributes to extract the language tag from.

-> Maybe LanguageTag
A language tag extracted from the given raw HTML attributes. If the given raw HTML attributes does not have lang attribute or its value is invalid, Nothing is returned.

Extracts the language tag from the given raw HTML attributes if it has lang attribute.

>>> extractLang ""
Nothing
>>> extractLang "lang=en"
Just "en"
>>> extractLang "lang=\"ko-KR\""
Just "ko-kr"
>>> extractLang " lang='ko-Hang'"
Just "ko-hang"

isKorean :: LanguageTag -> Bool

Determines whether the given language tag refers to any kind of Korean.

>>> isKorean "ko"
True
>>> isKorean "ko-KR"
True
>>> isKorean "kor-Hang"
True
>>> isKorean "en"
False
>>> isKorean "en-KR"
False

isNeverKorean :: Maybe LanguageTag -> Bool

Determines whether the given language tag undoubtedly does not refer to any kind of Korean.

>>> isNeverKorean $ Just "ko"
False
>>> isNeverKorean $ Just "ko-KR"
False
>>> isNeverKorean Nothing
False
>>> isNeverKorean $ Just "en"
True