Text.Seonbi.Html.Preservation

Documentation

isPreservedEntity :: HtmlEntity -> Bool

True if the given HTML entity should be preserved from transformation.

isPreservedTag :: HtmlTag -> Bool

True if the given tag should be preserved from transformation.

isPreservedTagStack :: HtmlTagStack -> Bool

True if the given tag stack should be preserved from transformation.