Text.Seonbi.Trie

Description

A trie from Text keys to values.

Documentation

data Trie a

A trie from Text keys to a values.

elems :: Trie a -> [a]

Lists all values in the trie. Values are ordered by their associated keys.

empty :: Trie a

The empty trie.

fromList :: [(Text, a)] -> Trie a

Converts a list of associated pairs into a trie. For duplicate keys, values earlier in the list shadow later ones.

insert

Arguments

:: Text
A new key to insert. If there is already the same key in the trie, the existing value is overwritten by the new value.

-> a
A value associated to the key.

-> Trie a
An existing trie.

-> Trie a
The new trie with the inserted key.

Inserts a new key into the trie.

keys :: Trie a -> [Text]

Lists all keys in the trie. Keys will be ordered.

lookup :: Text -> Trie a -> Maybe a

Gets the value associated with a key if it exists.

member :: Text -> Trie a -> Bool

Checks if a key has a value in the trie.

mergeBy :: (a -> a -> Maybe a) -> Trie a -> Trie a -> Trie a

Combines two tries, using a function to resolve collisions. This can only define the space of functions between union and symmetric difference but, with those two, all set operations can be defined (albeit inefficiently).

null :: Trie a -> Bool

Checks if the trie is empty.

singleton :: Text -> a -> Trie a

Constructs a singleton trie.

size :: Trie a -> Int

Gets the number of elements in the trie.

toList :: Trie a -> [(Text, a)]

Converts a trie into a list of associated pairs. Keys will be ordered.

unionL :: Trie a -> Trie a -> Trie a

Combines two tries, resolving conflicts by choosing the value from the left (former) trie.

unionR :: Trie a -> Trie a -> Trie a

Combines two tries, resolving conflicts by choosing the value from the right (latter) trie.