Text.Seonbi.Unihan.KHangul

Documentation

data CharacterSet

Represents character set standards for Korean writing system.

Constructors

KS_X_1001
KS X 1001 (정보 교환용 부호계).

KS_X_1002
KS X 1002 (정보 교환용 부호 확장 세트).

NonStandard
Represents that a Hanja character is not included in any Korean character set standards.

type HanjaReadings = Map Char HanjaReadingCitation

All readings of a Hanja character.

data HanjaReadingCitation

Represents what standard a reading of character belongs to and a purpose of the reading.

Constructors

HanjaReadingCitation CharacterSet (Set Purpose)

type KHangulData = Map Char HanjaReadings

Maps all Hanja characters to their possible readings.

data Purpose

Represents purposes of Hanja characters.

Constructors

Education
Basic Hanja for educational use (漢文敎育用基礎漢字), a subset of Hanja defined in 1972 by a South Korean standard for educational use.

PersonalName
Hanja for personal names (人名用漢字).

kHangulData :: KHangulData

Data that map Hanja characters to their corresponding kHangul entries (i.e., Hanja readings and citations).

>>> import Data.Map.Strict as M
>>> let Just entries = M.lookup '天' kHangulData
>>> entries
fromList [('천',HanjaReadingCitation KS_X_1001 (fromList [Education]))]

kHangulData' :: Either String KHangulData